avoid recursive import
