+++ src/index.test.ts
@@ -7,10 +7,9 @@ import { InExpression } from './expression/in'
 import { MathExpression } from './expression/math'
 import { Value } from './expression/value'
 import { Query } from './query'
-import { JoinClause } from './query/joinClause'
 import { OrderingTerm } from './query/orderingTerm'
 import { ResultColumn } from './query/resultColumn'
-import { JoinedTableOrSubquery } from './query/tableOrSubquery'
+import { JoinClause, JoinedTableOrSubquery } from './query/tableOrSubquery'
 
 test('SELECT * FROM Student', () => {
   const query = new Query({ $from: 'Student' })
