+++ src/index.ts
@@ -17,7 +17,6 @@ export { Unknown, IUnknown } from './expression/unknown'
 export { Value, IValue } from './expression/value'
 export { Query, IQuery } from './query'
 export { ResultColumn, IResultColumn } from './query/resultColumn'
-export { JoinOperator, JoinClause, IJoinClause } from './query/joinClause'
-export { TableOrSubquery, ITableOrSubquery, JoinedTableOrSubquery, IJoinedTableOrSubquery } from './query/tableOrSubquery'
+export { TableOrSubquery, ITableOrSubquery, JoinOperator, JoinClause, IJoinClause, JoinedTableOrSubquery, IJoinedTableOrSubquery } from './query/tableOrSubquery'
 export { GroupBy, IGroupBy } from './query/groupBy'
 export { Order, OrderingTerm, IOrderingTerm } from './query/orderingTerm'
